requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis Cache
redis==5.0.1

# AI/ML
openai==1.3.8
anthropic==0.7.8
langchain==0.0.350
huggingface-hub==0.19.4

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0

# Data Processing
pydantic==2.5.1
pydantic-settings==2.0.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0